Update README.md
Election-Audit Results
The analysis of the election show that:

There were 369,711 votes cast in the election

County Votes breakdown:

Jefferson: 10.5% (38,855)
Denver: 82.8% (306,055)
Arapahoe: 6.7% (24,801)
County wuth largest number of votes: Denver

The candidates were:

Charles Casper Stockham
Diana DeGette
Raymon Anthony Doane
Candidate Votes breakdown:

Charles Casper Stockham received 23.0% of the vote and 85,213 number of votes.
Diana DeGette received 73.8% of the vote and 272,892 number of votes.
Raymon Anthony Doane received 3.1% of the vote and 11,606 number of votes.
The winnner of the election was:

Diana DeGette, who received 73.8% of vote and 272,892 number of votes.
Election Audit Summary
This python script can be reused with minor changes for any election audits provided the structure of the election data is in csv format is not changed. The following tweaks may needs to be done inorder to reuse the program

Tweak for the csv file name and path of the file
Tweak may be required if the column number for the candidate name or county name is changed.